+++ Day.Four.CeresSearch/PartOne.cs
@@ -74,7 +74,7 @@ public class PartOne : Compute
         return count;
     }
 
-    protected override int Run()
+    protected override long Run()
     {
         int result = 0;
