+++ Day.Four.CeresSearch/PartTwo.cs
@@ -6,7 +6,7 @@ public class PartTwo : Compute
 {
     public PartTwo() : base("Day 4, Part 2") { }
 
-    protected override int Run()
+    protected override long Run()
     {
         int result = 0;
